removed unused, redundant cmd/version package
